altering the way fees are loaded

+++ fees/trade/filter/TradeFeeFilter.py
@@ -1,7 +1,10 @@
+from core.number.BigFloat import BigFloat
+
+
 class TradeFeeFilter:
 
-    def obtain_account_trade_fee(self) -> float:
+    def obtain_account_trade_fee(self) -> BigFloat:
         pass
 
-    def obtain_instrument_trade_fee(self, instrument) -> float:
+    def obtain_instrument_trade_fee(self, instrument) -> BigFloat:
         pass
